BugFix playerName
The problem was that the input text field shared the same onclick as the dialog, so the dialog would still continue as you were trying to input name. This is no longer the case. Now for the player you have to click inside the dialog box to get through the dialogue